improving running instructions and configure.bash

+++ configure.bash
@@ -1,18 +1,41 @@
 
-echo "Dowloading Love game engine..."
 
-rm -r dependencies 2> /dev/null
-mkdir dependencies 2> /dev/null
+os='unknown'
+uname_out=$(uname)
+if [[ "$uname_out" == "Linux" ]]; then
+   os='linux'
+elif [[ "$uname_out" == "Darwin" ]]; then
+   os="osx"
+fi
+
+
+if [ ! -d dependencies ]; then
+	echo "Dowloading Love game engine..."
+	mkdir dependencies
+fi
+
 cd dependencies
 
-wget https://bitbucket.org/rude/love/downloads/love-0.10.1-macosx-x64.zip
-tar -xf love-0.10.1-macosx-x64.zip
-rm *.zip
+if [ ! -d love.app ]; then
+
+	if [[ "$os" == "osx" ]]; then
 
 
-echo "Setting up alias for engine application..."
+		wget https://bitbucket.org/rude/love/downloads/love-0.10.1-macosx-x64.zip
+		tar -xf love-0.10.1-macosx-x64.zip
+		rm *.zip
 
-LOVE_PATH=$(pwd)
-alias love=$LOVE_PATH/love.app/Contents/MacOS/love
+		echo "Setting up alias for engine application..."
+		LOVE_PATH=$(pwd)
+
+		alias love=$LOVE_PATH/love.app/Contents/MacOS/love
+		#echo "alias love=$LOVE_PATH/love.app/Contents/MacOS/love" >> $HOME/.profile
+	elif [[ "$os" == "linux" ]]; then
+		echo "TODO"
+	fi
+
+fi
 
 cd ..
+
+
